learningdoc/requirements.txt
langchain == 0.2.12
langchain-openai == 0.1.20
langchain-community == 0.2.11
python-dotenv == 1.0.1
faiss-cpu == 1.8.0.post1
pdfminer.six == 20240706
wikipedia == 1.4.0